+++ sample-apps/segmentation_covid_crf/main.py
@@ -16,7 +16,6 @@ from monailabel.utils.infer.deepgrow_3d import InferDeepgrow3D
 from monailabel.utils.others.post import BoundingBoxd, Restored
 
 from lib import MyActiveLearning, MyCRF, MyInfer, MyTrain
-from lib.transforms import AddUnaryTermd, ApplyCRFPostProcd
 
 logger = logging.getLogger(__name__)
