+++ sample-apps/segmentation_spleen_crf/lib/postproc_crf.py
@@ -1,6 +1,6 @@
 import logging
 from monailabel.utils.postproc import PostProcessingTask, PostProcType
-from lib.transforms import AddUnaryTermd, ApplyCRFPostProcd
+from monailabel.utils.postproc.postproctx import AddUnaryTermd, ApplyMONAICRFPostProcd
 
 from monai.transforms import (
     LoadImaged,
@@ -46,11 +46,11 @@ class MyCRF(PostProcessingTask):
             
             # at the moment CRF implementation is bottleneck taking a long time, 
             # therefore scaling non-isotropic with big spacing
-            Spacingd(keys=['image', 'logits'], pixdim=[1.0, 1.0, 5.0]),
-            Spacingd(keys=['scribbles'], pixdim=[1.0, 1.0, 5.0], mode='nearest'),
+            Spacingd(keys=['image', 'logits'], pixdim=[2.0, 2.0, 5.0]),
+            Spacingd(keys=['scribbles'], pixdim=[2.0, 2.0, 5.0], mode='nearest'),
 
             ScaleIntensityRanged(keys='image', a_min=-164, a_max=164, b_min=0.0, b_max=1.0, clip=True),
-            AddUnaryTermd(ref_prob='logits', unary="unary", scribbles="scribbles", channel_dim=0, sc_background_label=2, sc_foreground_label=3, scale_infty=100, use_simplecrf=False),
+            AddUnaryTermd(ref_prob='logits', unary="unary", scribbles="scribbles", channel_dim=0, scribbles_bg_label=2, scribbles_fg_label=3, scale_infty=100, use_simplecrf=False),
             AddChanneld(keys=['image', 'unary']),
             ToTensord(keys=['image', 'logits', 'unary'])
 
@@ -72,5 +72,5 @@ class MyCRF(PostProcessingTask):
                 return ApplyCRFPostProcd(unary='unary', pairwise='image', post_proc_label='pred', use_simplecrf=False)
 
         """
-        return ApplyCRFPostProcd(unary='unary', pairwise='image', post_proc_label='pred', use_simplecrf=False)
+        return ApplyMONAICRFPostProcd(unary='unary', pairwise='image', post_proc_label='pred')
